Use pkg-config to find the gssapi system library
This uses pkg-config to locate the gssapi system library and the
necessary include paths.  This can be particularly useful when gssapi
has interface dependencies (mit has comerr includes in it's public
interface, for example) that are installed in different prefixes as is
often the case with third party package managers.